// src/app/teams/models/team.model.ts
export interface TeamModel {
    ID: number;
    image: string;
    fullname: string;
    fullname_url: string;
    grade: string;
    content: string;
    scan_doc: string;
    facebook: string;
    comptex: string;
    linkedin: string;
    counter: number;
    is_publie: boolean;
    signature: string;
    CreatedAt: Date;
    UpdatedAt: Date;
}